refactor(Bundles): move bundles to Bundles.scala
