+++ src/main/scala/xiangshan/mem/lsqueue/LSQWrapper.scala
@@ -23,14 +23,14 @@ import utils._
 import utility._
 import xiangshan._
 import xiangshan.backend.Bundles.{DynInst, MemExuOutput}
+import xiangshan.backend._
+import xiangshan.backend.rob.RobLsqIO
+import xiangshan.backend.fu.FuType
+import xiangshan.mem.Bundles._
 import xiangshan.cache._
 import xiangshan.cache.{DCacheWordIO, DCacheLineIO, MemoryOpConstants}
 import xiangshan.cache.{CMOReq, CMOResp}
 import xiangshan.cache.mmu.{TlbRequestIO, TlbHintIO}
-import xiangshan.mem._
-import xiangshan.backend._
-import xiangshan.backend.rob.RobLsqIO
-import xiangshan.backend.fu.FuType
 
 class ExceptionAddrIO(implicit p: Parameters) extends XSBundle {
   val isStore = Input(Bool())
@@ -248,7 +248,7 @@ class LsqWrapper(implicit p: Parameters) extends XSModule with HasDCacheParamete
   switch(pendingstate){
     is(s_idle){
       when(io.uncache.req.fire){
-        pendingstate := 
+        pendingstate :=
           Mux(io.uncacheOutstanding && io.uncache.req.bits.nc, s_idle,
           Mux(loadQueue.io.uncache.req.valid, s_load,
           s_store))
