+++ src/main/scala/xiangshan/mem/lsqueue/LoadQueueRAR.scala
@@ -15,15 +15,16 @@
 ***************************************************************************************/
 package xiangshan.mem
 
+import org.chipsalliance.cde.config._
 import chisel3._
 import chisel3.util._
-import org.chipsalliance.cde.config._
-import xiangshan._
-import xiangshan.backend.rob.RobPtr
-import xiangshan.cache._
 import utils._
 import utility._
+import xiangshan._
+import xiangshan.backend.rob.RobPtr
 import xiangshan.backend.Bundles.DynInst
+import xiangshan.mem.Bundles._
+import xiangshan.cache._
 
 class LoadQueueRAR(implicit p: Parameters) extends XSModule
   with HasDCacheParameters
