+++ src/main/scala/xiangshan/mem/prefetch/L1PrefetchComponent.scala
@@ -1,18 +1,19 @@
 package xiangshan.mem.prefetch
 
 import org.chipsalliance.cde.config.Parameters
-import freechips.rocketchip.util._
 import chisel3._
 import chisel3.util._
-import xiangshan._
+import freechips.rocketchip.util._
 import utils._
 import utility._
+import xiangshan._
 import xiangshan.backend.fu.PMPRespBundle
-import xiangshan.cache.HasDCacheParameters
-import xiangshan.cache.mmu._
-import xiangshan.mem.{L1PrefetchReq, LdPrefetchTrainBundle}
+import xiangshan.mem.L1PrefetchReq
+import xiangshan.mem.Bundles.{LdPrefetchTrainBundle}
 import xiangshan.mem.trace._
 import xiangshan.mem.L1PrefetchSource
+import xiangshan.cache.HasDCacheParameters
+import xiangshan.cache.mmu._
 
 trait HasL1PrefetchHelper extends HasCircularQueuePtrHelper with HasDCacheParameters {
   // region related
@@ -592,7 +593,7 @@ class MutiLevelPrefetchFilter(implicit p: Parameters) extends XSModule with HasL
   val s1_tlb_fire_vec = GatedValidRegNext(s0_tlb_fire_vec)
   val s2_tlb_fire_vec = GatedValidRegNext(s1_tlb_fire_vec)
   val s3_tlb_fire_vec = GatedValidRegNext(s2_tlb_fire_vec)
-  val not_tlbing_vec = VecInit((0 until MLP_SIZE).map{case i => 
+  val not_tlbing_vec = VecInit((0 until MLP_SIZE).map{case i =>
     !s1_tlb_fire_vec(i) && !s2_tlb_fire_vec(i) && !s3_tlb_fire_vec(i)
   })
 
@@ -673,7 +674,7 @@ class MutiLevelPrefetchFilter(implicit p: Parameters) extends XSModule with HasL
     }.otherwise {
       val inner_index = s3_tlb_update_index - MLP_L1_SIZE.U
       l2_array(inner_index).is_vaddr := s3_tlb_resp.miss
-  
+
       when(!s3_tlb_resp.miss) {
         l2_array(inner_index).region := Cat(0.U((VAddrBits - PAddrBits).W), s3_tlb_resp.paddr.head(s3_tlb_resp.paddr.head.getWidth - 1, REGION_TAG_OFFSET))
         when(s3_drop) {
@@ -833,7 +834,7 @@ class MutiLevelPrefetchFilter(implicit p: Parameters) extends XSModule with HasL
 
   XSPerfAccumulate("l2_prefetche_queue_busby", io.l2PfqBusy)
   XSPerfHistogram("filter_active", PopCount(VecInit(
-    l1_array.zip(l1_valids).map{ case (e, v) => e.can_send_pf(v) } ++ 
+    l1_array.zip(l1_valids).map{ case (e, v) => e.can_send_pf(v) } ++
     l2_array.zip(l2_valids).map{ case (e, v) => e.can_send_pf(v) }
     ).asUInt), true.B, 0, MLP_SIZE, 1)
   XSPerfHistogram("l1_filter_active", PopCount(VecInit(l1_array.zip(l1_valids).map{ case (e, v) => e.can_send_pf(v)}).asUInt), true.B, 0, MLP_L1_SIZE, 1)
