+++ src/main/scala/xiangshan/mem/prefetch/L1StreamPrefetcher.scala
@@ -3,14 +3,15 @@ package xiangshan.mem.prefetch
 import org.chipsalliance.cde.config.Parameters
 import chisel3._
 import chisel3.util._
-import xiangshan._
 import utils._
 import utility._
-import xiangshan.cache.HasDCacheParameters
-import xiangshan.cache.mmu._
-import xiangshan.mem.{L1PrefetchReq, LdPrefetchTrainBundle}
+import xiangshan._
+import xiangshan.mem.L1PrefetchReq
+import xiangshan.mem.Bundles.{LdPrefetchTrainBundle}
 import xiangshan.mem.trace._
 import xiangshan.mem.L1PrefetchSource
+import xiangshan.cache.HasDCacheParameters
+import xiangshan.cache.mmu._
 
 trait HasStreamPrefetchHelper extends HasL1PrefetchHelper {
   // capacity related
@@ -85,7 +86,7 @@ class StreamBitVectorBundle(implicit p: Parameters) extends XSBundle with HasStr
     }else {
       decr_mode := INIT_DEC_MODE.B
     }
-    
+
 
     assert(PopCount(alloc_bit_vec) === 1.U, "alloc vector should be one hot")
   }
@@ -126,7 +127,7 @@ class StreamPrefetchReqBundle(implicit p: Parameters) extends XSBundle with HasS
     res.source.value := source
 
     res.trigger_pc := t_pc
-    res.trigger_va := t_va 
+    res.trigger_va := t_va
 
     val region_bits = get_region_bits(vaddr)
     val region_bit_vec = UIntToOH(region_bits)
