+++ src/main/scala/xiangshan/mem/prefetch/L1StridePrefetcher.scala
@@ -27,13 +27,14 @@ package xiangshan.mem.prefetch
 import org.chipsalliance.cde.config.Parameters
 import chisel3._
 import chisel3.util._
-import xiangshan._
 import utils._
 import utility._
+import xiangshan._
+import xiangshan.mem.L1PrefetchReq
+import xiangshan.mem.Bundles.{LdPrefetchTrainBundle}
+import xiangshan.mem.trace._
 import xiangshan.cache.HasDCacheParameters
 import xiangshan.cache.mmu._
-import xiangshan.mem.{L1PrefetchReq, LdPrefetchTrainBundle}
-import xiangshan.mem.trace._
 import scala.collection.SeqLike
 
 trait HasStridePrefetchHelper extends HasL1PrefetchHelper {
