+++ src/main/scala/xiangshan/mem/prefetch/PrefetcherMonitor.scala
@@ -1,12 +1,13 @@
 package xiangshan.mem.prefetch
 
+import org.chipsalliance.cde.config.Parameters
 import chisel3._
 import chisel3.util._
-import org.chipsalliance.cde.config.Parameters
-import xiangshan._
-import xiangshan.mem.{LdPrefetchTrainBundle, StPrefetchTrainBundle, L1PrefetchReq}
 import utils._
 import utility._
+import xiangshan._
+import xiangshan.mem.L1PrefetchReq
+import xiangshan.mem.Bundles.{LdPrefetchTrainBundle, StPrefetchTrainBundle}
 
 trait HasPrefetcherMonitorHelper {
   val TIMELY_CHECK_INTERVAL = 1000
